+++ PGMutagen/src/PGMutagenWrapper.cpp
@@ -172,6 +172,7 @@ auto PGMutagenWrapper::libGetModelUses(const std::wstring& modelPath) -> std::ve
         curUse.subModel = string(mu->sub_model()->begin(), mu->sub_model()->end());
         curUse.isWeighted = mu->is_weighted();
         curUse.singlepassMATO = mu->singlepass_mato();
+        curUse.isIgnored = mu->is_ignored();
 
         for (const auto* const altTex : *mu->alternate_textures()) {
             auto curAltTex = AlternateTexture();
@@ -237,7 +238,7 @@ void PGMutagenWrapper::libSetModelUses(const std::vector<ModelUse>& modelUses)
         const auto altTexVectorOffset = builder.CreateVector(altTexOffsets);
 
         const auto modelUseOffset = PGMutagenBuffers::CreateModelUse(
-            builder, modNameOffset, mu.formID, subModelOffset, false, meshFileOffset, false, altTexVectorOffset);
+            builder, modNameOffset, mu.formID, subModelOffset, false, meshFileOffset, false, false, altTexVectorOffset);
         modelUsesOffsets.push_back(modelUseOffset);
     }